Refactored parse util functions out of ion_loader
